move jquery and vue file
