+++ example/example2/js/index.js
@@ -1,4 +1,5 @@
 import '../css/app.css'
+import '../../../jquery'
 
 $.parallax({ direction: 'x' })
