+++ example/example3/js/index.js
@@ -1,1 +1,2 @@
 import '../css/app.css'
+import '../../../jquery'
