+++ example/example5/js/index.js
@@ -1,5 +1,6 @@
 import Vue from 'vue'
-import Parallax from '../../js/scrollParallax'
+import Parallax from '../../../vue'
+
 import App from './App'
 
 Vue.use(Parallax, { direction: 'x'})
