+++ public/example1/index.html
@@ -144,7 +144,6 @@
 </main>
 
 <script type="text/javascript" src="../js/jquery.js"></script>
-<script type="text/javascript" src="../js/jquery.scrollParallax.js"></script>
 <script type="text/javascript" src="./js/app.js"></script>
 </body>
 </html>
