+++ public/example2/index.html
@@ -180,7 +180,6 @@
 </main>
 
 <script type="text/javascript" src="../js/jquery.js"></script>
-<script type="text/javascript" src="../js/jquery.scrollParallax.js"></script>
 <script type="text/javascript" src="./js/app.js"></script>
 </body>
 </html>
