+++ public/example3/index.html
@@ -83,7 +83,7 @@
 
 <script src="../js/jquery.js"></script>
 <script src="./js/jquery.dataExtend.js"></script>
-<script src="../js/jquery.scrollParallax.js"></script>
+<script src="./js/app.js"></script>
 
 
 <script>
